whole files have been modified

+++ quackapp-backend/src/main/java/de/quackr/api/UserController.java
@@ -10,6 +10,7 @@ import java.util.regex.PatternSyntaxException;
 import javax.ejb.EJB;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.print.attribute.standard.Media;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
@@ -21,10 +22,16 @@ import de.quackr.persistence.entities.Quack;
 
 import de.quackr.persistence.entities.User;
 import de.quackr.persistence.entities.User_;
+import org.apache.shiro.SecurityUtils;
+import org.apache.shiro.subject.Subject;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import static de.quackr.auth.DefaultRealm.ROLE_ADMIN;
 
 /**
  * @author wangxiaoshi
- * 
+ *
  */
 @Path("/users")
 public class UserController {
@@ -33,6 +40,8 @@ public class UserController {
     static final String REGEX_PASSWORD = "^[a-zA-Z0-9]{6,16}$";
     static final String REGEX_EMAIL = "^([a-z0-9A-Z]+[-|\\.]?)+[a-z0-9A-Z]@([a-z0-9A-Z]+(-[a-z0-9A-Z]+)?\\.)+[a-zA-Z]{2,}$";
 
+    private Logger logger = LoggerFactory.getLogger(UserController.class);
+
     @EJB
     private IOController ioController;
 
@@ -40,14 +49,48 @@ public class UserController {
     @Path("/{id}")
     @Produces(MediaType.APPLICATION_JSON)
     public Response readUser(@PathParam("id") long id) {
-        return Response.ok(ioController.getUser(id)).build();
+        Subject subject = SecurityUtils.getSubject();
+
+        if (!subject.isAuthenticated()) {
+            return Response.status(Response.Status.UNAUTHORIZED).build();
+        }
+
+        User user = ioController.getUser(id);
+
+        if (!user.getUsername().equals(subject.getPrincipal()) && !user.isModerator() && !user.isAdmin()) {
+            return Response.status(Response.Status.UNAUTHORIZED).build();
+        }
+
+        return Response.ok(user).build();
     }
 
     @GET
-    @Path("/all")
+    @Path("/me")
     @Produces(MediaType.APPLICATION_JSON)
-    public List<User> readAllUsersAsJSON() {
-        return ioController.getAllUsers();
+    public Response readAuthenticatedUser() {
+        Subject subject = SecurityUtils.getSubject();
+
+        if (!subject.isAuthenticated()) {
+            return Response.status(Response.Status.UNAUTHORIZED).build();
+        }
+
+        User user = ioController.getUser((String) subject.getPrincipal());
+        return Response.ok(user).build();
+    }
+
+    @GET
+    @Produces(MediaType.APPLICATION_JSON)
+    public Response readAllUsersAsJSON() {
+        Subject subject = SecurityUtils.getSubject();
+
+        if(subject.isAuthenticated() && subject.hasRole(ROLE_ADMIN)) {
+            logger.debug("User \"{}\" is authenticated and admin. Returning all users.", subject.getPrincipal());
+            List<User> users = ioController.getAllUsers();
+            return Response.ok(users).build();
+        } else {
+            logger.debug("User is not authenticated or does not have admin role.");
+            return Response.status(Response.Status.UNAUTHORIZED).build();
+        }
     }
 
     @POST
@@ -56,10 +99,10 @@ public class UserController {
     public Response createUser(User param) {
         try {
             if(!validateInput(param)) {
-                Response.status(Response.Status.BAD_REQUEST).build();
+                return Response.status(Response.Status.BAD_REQUEST).build();
             }
         } catch (PatternSyntaxException ex) {
-            Response.serverError();
+            return Response.serverError().build();
         }
 
         final User user = new User();
@@ -81,24 +124,44 @@ public class UserController {
     @DELETE
     @Path("/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public void deleteUser(@PathParam("id") long id) {
+    public Response deleteUser(@PathParam("id") long id) {
+        if (!userLoggedInOrAdmin(id)) {
+            return Response.status(Response.Status.UNAUTHORIZED).build();
+        }
+
+        logger.info("Deleting user with ID #{}.", id);
         ioController.deleteUser(id);
+
+        return Response.ok().build();
     }
 
     @PUT
     @Path("/{id}")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public void updateUser(@PathParam("id") long id, User param) {
+    public Response updateUser(@PathParam("id") long id, User param) {
+        logger.debug("Trying to update user with ID #{}.", id);
+
+        if (!userLoggedInOrAdmin(id)) {
+            return Response.status(Response.Status.UNAUTHORIZED).build();
+        }
+
         try {
-            if(!validateInput(param)) {
-                Response.status(Response.Status.BAD_REQUEST).build();
+            if(param.getUsername() != null && !validateUsername(param.getUsername())) {
+                return Response.status(Response.Status.BAD_REQUEST).build();
+            }
+            if(param.getPasswordHash() != null && !validatePassword(param.getPasswordHash())) {
+                return Response.status(Response.Status.BAD_REQUEST).build();
+            }
+            if(param.getEmail() != null && !validateEmail(param.getEmail())) {
+                return Response.status(Response.Status.BAD_REQUEST).build();
             }
         } catch (PatternSyntaxException ex) {
-            Response.serverError();
+            return Response.serverError().build();
         }
 
         ioController.updateUser(id, param);
+        return Response.ok(param).build();
     }
 
     public static boolean validateInput(User param) {
@@ -117,6 +180,18 @@ public class UserController {
 
     public static boolean validateEmail(String email) {
         return Pattern.matches(REGEX_EMAIL, email);
+    }
+
+    private boolean userLoggedInOrAdmin(long id) {
+        Subject subject = SecurityUtils.getSubject();
+
+        if (!subject.isAuthenticated()) {
+            return false;
+        }
+
+        User user = ioController.getUser(id);
+
+        return user.getId() == id || user.isAdmin();
     }
 
 }
